requirements.txt
aiohttp>=3.6,<4.0
aiosignal==1.3.1
anaconda-client==1.7.2
anaconda-navigator==2.0.3
anndata==0.9.1
annotated-types==0.6.0
annoy==1.17.3
anyio==4.3.0
appdirs==1.4.4
applaunchservices==0.2.1
argh==0.26.2
async-timeout==4.0.3
atomicwrites==1.4.0
awsebcli==3.20.10
backports.weakref==1.0.post1
bkcharts==0.2
black==19.10b0
blessed==1.20.0
blinker==1.7.0
boto==2.49.0
botocore==1.31.85
Bottleneck==1.3.2
brotlipy==0.7.0
bs4==0.0.2
cachelib==0.12.0
celltypist==1.5.2
cement==2.8.2
certifi==2024.2.2
cfgv==3.3.1
charset-normalizer==3.3.2
click==8.1.7
clyent==1.2.2
colorama==0.4.3
conda==4.10.1
conda-build==3.21.4
conda-verify==3.4.2
contextlib2==0.6.0.post1
contourpy==1.1.0
cycler==0.10.0
cytoolz==0.11.0
dataclasses-json==0.6.4
distlib==0.3.6
distro==1.9.0
dominate==2.7.0
entrypoints==0.3
et-xmlfile==1.0.1
exceptiongroup==1.2.0
fastcache==1.1.0
filelock==3.12.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Session==0.8.0
fonttools==4.40.0
frozenlist==1.4.1
funcy==2.0
future==0.18.2
fuzzywuzzy==0.18.0
gensim==4.3.2
gmpy2==2.0.8
greenlet==2.0.2
h11==0.14.0
h5py==3.9.0
HeapDict==1.0.1
httpcore==1.0.5
httpx==0.27.0
identify==2.5.7
igraph==0.10.5
importlib-resources==5.12.0
ipycanvas==0.9.0
itsdangerous==2.1.2
jdcal==1.4.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json5==0.9.5
jsonpatch==1.33
jsonpointer==2.4
jupyter==1.0.0
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.40
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.39
leidenalg==0.10.0
Levenshtein==0.21.1
llvmlite==0.36.0
load-dotenv==0.1.0
locket==0.2.1
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.7.1
mccabe==0.6.1
mkl-fft==1.3.0
mkl-service==2.3.0
mpmath==1.2.1
msgspec==0.18.6
multidict==6.0.5
multipledispatch==0.6.0
mypy-extensions==0.4.3
natsort==8.4.0
navigator-updater==0.2.1
nodeenv==1.7.0
numpy==1.24.4
olefile==0.46
openai==1.16.1
ordered-set==4.1.0
orjson==3.10.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
parso==0.7.0
pathspec==0.10.1
patsy==0.5.1
penngrader==0.4
pep8==1.7.1
pipenv==2023.4.29
pkginfo==1.7.0
platformdirs==3.5.1
playwright==1.38.0
ply==3.11
pre-commit==2.20.0
pubplot==0.2.4
pycosat==0.6.3
pycurl==7.43.0.6
pydantic==2.6.4
pydantic_core==2.16.3
pyee==9.0.4
PyLaTeX==1.4.1
pyLDAvis==3.4.0
PyMySQL==1.1.0
pynndescent==0.5.10
pyodbc===4.0.0-unsupported
pytest==6.2.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-Levenshtein==0.21.1
PyYAML==5.4.1
pyzmq==20.0.0
QDarkStyle==2.8.1
QtPy==1.9.0
rapidfuzz==3.3.0
regex==2023.12.25
requests==2.31.0
scanpy==1.9.3
scikit-image==0.18.1
scikit-learn==1.3.1
scipy==1.10.1
selenium==4.18.1
semantic-version==2.8.5
session-info==1.0.0
simplegeneric==0.8.1
smart-open==6.4.0
sniffio==1.3.0
stdlib-list==0.9.0
tables==3.6.1
tenacity==8.2.3
termcolor==1.1.0
terminado==0.9.4
textblob==0.17.1
texttable==1.6.7
tifffile==2020.10.1
tiktoken==0.6.0
torch==1.13.1
torchvision==0.14.1
trio==0.24.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
umap-learn==0.5.3
unicodecsv==0.14.1
urllib3==1.26.18
virtualenv==20.23.0
virtualenv-clone==0.5.7
watchdog==4.0.0
wcwidth==0.1.9
webdriver-manager==4.0.1
webencodings==0.5.1
Werkzeug==3.0.2
widgetsnbextension==3.5.1
wordcloud==1.9.2
wsproto==1.2.0
xlwings==0.23.0
xlwt==1.3.0
yarl==1.9.4
zict==2.0.0
zope.event==4.5.0
